refactor: change the UsersList to UsersTable
