+++ src/components/UsersTable.tsx
@@ -1,0 +1,31 @@
+type User = {
+  name: string;
+  email: string;
+};
+
+type UserListProps = {
+  users: User[];
+};
+
+const UsersList: React.FC<UserListProps> = ({ users }) => {
+  return (
+    <table className='users-table'>
+      <thead>
+        <tr>
+          <th>Nome</th>
+          <th>Email</th>
+        </tr>
+      </thead>
+      <tbody>
+        {users.map((user, index) => (
+          <tr key={index}>
+            <td>{user.name}</td>
+            <td>{user.email}</td>
+          </tr>
+        ))}
+      </tbody>
+    </table>
+  );
+};
+
+export default UsersList;
