+++ src/pages/Users.tsx
@@ -1,4 +1,4 @@
-import UsersList from '../components/UsersList';
+import UsersList from '../components/UsersTable';
 
 const Users = () => {
   const mockData = {
